LPS-77727 Move search classes to dynamic-data-mapping-web and apply usages
