+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-web/src/main/java/com/liferay/dynamic/data/mapping/web/internal/search/StructureDisplayTerms.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.search;
+package com.liferay.dynamic.data.mapping.web.internal.search;
 
 import com.liferay.portal.kernel.dao.search.DisplayTerms;
 import com.liferay.portal.kernel.util.ParamUtil;
@@ -22,20 +22,24 @@ import javax.portlet.PortletRequest;
 /**
  * @author Eduardo Lundgren
  */
-public class TemplateDisplayTerms extends DisplayTerms {
+public class StructureDisplayTerms extends DisplayTerms {
 
 	public static final String DESCRIPTION = "description";
 
 	public static final String NAME = "name";
 
-	public static final String TYPE = "type";
+	public static final String SEARCH_RESTRICTION = "searchRestriction";
 
-	public TemplateDisplayTerms(PortletRequest portletRequest) {
+	public static final String STORAGE_TYPE = "storageType";
+
+	public StructureDisplayTerms(PortletRequest portletRequest) {
 		super(portletRequest);
 
 		description = ParamUtil.getString(portletRequest, DESCRIPTION);
 		name = ParamUtil.getString(portletRequest, NAME);
-		type = ParamUtil.getString(portletRequest, TYPE);
+		searchRestriction = ParamUtil.getBoolean(
+			portletRequest, SEARCH_RESTRICTION);
+		storageType = ParamUtil.getString(portletRequest, STORAGE_TYPE);
 	}
 
 	public String getDescription() {
@@ -46,12 +50,17 @@ public class TemplateDisplayTerms extends DisplayTerms {
 		return name;
 	}
 
-	public String getType() {
-		return type;
+	public String getStorageType() {
+		return storageType;
+	}
+
+	public Boolean isSearchRestriction() {
+		return searchRestriction;
 	}
 
 	protected String description;
 	protected String name;
-	protected String type;
+	protected Boolean searchRestriction;
+	protected String storageType;
 
 }
