+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-web/src/main/resources/META-INF/resources/init.jsp
@@ -48,12 +48,6 @@ page import="com.liferay.dynamic.data.mapping.model.DDMStructureVersion" %><%@
 page import="com.liferay.dynamic.data.mapping.model.DDMTemplate" %><%@
 page import="com.liferay.dynamic.data.mapping.model.DDMTemplateConstants" %><%@
 page import="com.liferay.dynamic.data.mapping.model.DDMTemplateVersion" %><%@
-page import="com.liferay.dynamic.data.mapping.search.StructureDisplayTerms" %><%@
-page import="com.liferay.dynamic.data.mapping.search.StructureSearch" %><%@
-page import="com.liferay.dynamic.data.mapping.search.StructureSearchTerms" %><%@
-page import="com.liferay.dynamic.data.mapping.search.TemplateDisplayTerms" %><%@
-page import="com.liferay.dynamic.data.mapping.search.TemplateSearch" %><%@
-page import="com.liferay.dynamic.data.mapping.search.TemplateSearchTerms" %><%@
 page import="com.liferay.dynamic.data.mapping.service.DDMStorageLinkLocalServiceUtil" %><%@
 page import="com.liferay.dynamic.data.mapping.service.DDMStructureLinkLocalServiceUtil" %><%@
 page import="com.liferay.dynamic.data.mapping.service.DDMStructureLocalServiceUtil" %><%@
@@ -75,6 +69,12 @@ page import="com.liferay.dynamic.data.mapping.web.configuration.DDMWebConfigurat
 page import="com.liferay.dynamic.data.mapping.web.internal.display.context.DDMDisplayContext" %><%@
 page import="com.liferay.dynamic.data.mapping.web.internal.search.DDMStructureRowChecker" %><%@
 page import="com.liferay.dynamic.data.mapping.web.internal.search.DDMTemplateRowChecker" %><%@
+page import="com.liferay.dynamic.data.mapping.web.internal.search.StructureDisplayTerms" %><%@
+page import="com.liferay.dynamic.data.mapping.web.internal.search.StructureSearch" %><%@
+page import="com.liferay.dynamic.data.mapping.web.internal.search.StructureSearchTerms" %><%@
+page import="com.liferay.dynamic.data.mapping.web.internal.search.TemplateDisplayTerms" %><%@
+page import="com.liferay.dynamic.data.mapping.web.internal.search.TemplateSearch" %><%@
+page import="com.liferay.dynamic.data.mapping.web.internal.search.TemplateSearchTerms" %><%@
 page import="com.liferay.petra.string.CharPool" %><%@
 page import="com.liferay.petra.string.StringPool" %><%@
 page import="com.liferay.portal.kernel.bean.BeanParamUtil" %><%@
